refactor(intents): consolidate all handlers in handlers.go
Move all handle* functions from helpers.go and intent.go to handlers.go
per new Handler Organization Rules added to AGENTS.md.

- handlers.go now contains ALL handle* functions (~700 lines)
- helpers.go contains only utilities (NO handle* functions)
- intent.go contains only Init/Update/View (~110 lines)
- AGENTS.md updated with Handler Organization Rules (MANDATORY)

AI-Generated-By: OpenCode (Claude Sonnet 4)
