+++ internal/cli/intents/burst_management/handlers.go
@@ -1,11 +1,14 @@
 package burst_management
 
 import (
+	"context"
+	"errors"
 	"fmt"
 
 	"github.com/baphled/kariya/internal/cli/behaviors"
 	"github.com/baphled/kariya/internal/cli/screens"
 	burstscreens "github.com/baphled/kariya/internal/cli/screens/burst_management"
+	burstmodals "github.com/baphled/kariya/internal/cli/screens/burst_management/modals"
 	"github.com/baphled/kariya/internal/cli/screens/facts"
 	"github.com/baphled/kariya/internal/cli/screens/timeline"
 	"github.com/baphled/kariya/internal/cli/uikit/feedback"
@@ -15,6 +18,13 @@ import (
 
 // Ensure Intent implements ScreenResultHandler interface.
 var _ behaviors.ScreenResultHandler = (*Intent)(nil)
+
+// noopCmd is a command that does nothing but prevents message propagation.
+func noopCmd() tea.Msg { return nil }
+
+// =============================================================================
+// Screen Result Handlers (ScreenResultHandler interface)
+// =============================================================================
 
 // handleScreenResult processes a screen result and determines next action.
 func (i *Intent) handleScreenResult(result interface{}) tea.Cmd {
@@ -103,6 +113,10 @@ func (i *Intent) HandleError(result *screens.ErrorResult) tea.Cmd {
 	return nil
 }
 
+// =============================================================================
+// Action Data Handler
+// =============================================================================
+
 // handleActionData processes action data from navigation results.
 //
 //nolint:funlen // Action routing requires handling multiple action types in one function
@@ -189,4 +203,501 @@ func (i *Intent) handleActionData(actionData map[string]interface{}) tea.Cmd {
 	default:
 		return nil
 	}
+}
+
+// =============================================================================
+// Modal Update Handlers
+// =============================================================================
+
+// handleModalUpdates handles updates for all modals in priority order.
+// Returns a command (possibly noopCmd) if a modal consumed the message.
+func (i *Intent) handleModalUpdates(msg tea.Msg) tea.Cmd {
+	// Process modals in priority order - first match handles the message.
+	if cmd := i.handleErrorModalUpdate(msg); cmd != nil {
+		return cmd
+	}
+	if cmd := i.handleLoadingModalUpdate(msg); cmd != nil {
+		return cmd
+	}
+	if cmd := i.handleDeleteModalUpdate(msg); cmd != nil {
+		return cmd
+	}
+	if cmd := i.handleConfirmModalUpdate(msg); cmd != nil {
+		return cmd
+	}
+	if cmd := i.handleSuggestionModalUpdate(msg); cmd != nil {
+		return cmd
+	}
+	if cmd := i.handleDetailModalUpdate(msg); cmd != nil {
+		return cmd
+	}
+	if cmd := i.handleEventsModalUpdate(msg); cmd != nil {
+		return cmd
+	}
+	if cmd := i.handleFactsModalUpdate(msg); cmd != nil {
+		return cmd
+	}
+	if cmd := i.handleEditModalUpdate(msg); cmd != nil {
+		return cmd
+	}
+	return nil
+}
+
+// handleErrorModalUpdate handles error modal updates (highest priority).
+func (i *Intent) handleErrorModalUpdate(msg tea.Msg) tea.Cmd {
+	if i.errorModal == nil {
+		return nil
+	}
+	if keyMsg, ok := msg.(tea.KeyMsg); ok && keyMsg.Type == tea.KeyEsc {
+		i.errorModal = nil
+	}
+	return noopCmd
+}
+
+// handleLoadingModalUpdate handles loading modal updates (cancellable with Esc).
+// Only forwards spinner tick messages to the modal; other messages are consumed
+// without restarting the spinner to avoid multiple concurrent tick loops.
+func (i *Intent) handleLoadingModalUpdate(msg tea.Msg) tea.Cmd {
+	if i.loadingModal == nil {
+		return nil
+	}
+	switch msg := msg.(type) {
+	case tea.KeyMsg:
+		if msg.Type == tea.KeyEsc {
+			i.cancelAsyncOperation()
+			return noopCmd
+		}
+		// Consume other keys without restarting spinner ticks.
+		return noopCmd
+	case feedback.ModalSpinnerTickMsg:
+		// Forward tick to loading modal to advance spinner.
+		return i.loadingModal.Update(msg)
+	default:
+		// Ignore unrelated messages while loading modal is active.
+		return noopCmd
+	}
+}
+
+// cancelAsyncOperation cancels the current async operation and resets state.
+func (i *Intent) cancelAsyncOperation() {
+	if i.cancelFunc != nil {
+		i.cancelFunc()
+		i.cancelFunc = nil
+	}
+	i.loadingModal = nil
+	i.suggestionsLoading = false
+	i.extractingFacts = false
+	i.state = StateList
+}
+
+// handleDeleteModalUpdate handles delete confirmation modal updates.
+func (i *Intent) handleDeleteModalUpdate(msg tea.Msg) tea.Cmd {
+	if i.deleteModal == nil || !i.deleteModal.IsVisible() {
+		return nil
+	}
+	cmd, confirmed := i.deleteModal.Update(msg)
+	if !i.deleteModal.IsVisible() {
+		if confirmed && i.selectedBurst != nil {
+			i.deleteBurst(i.selectedBurst)
+			return cmd
+		}
+		i.deleteModal = nil
+		i.selectedBurst = nil
+		i.state = StateList
+		return noopCmd
+	}
+	return cmd
+}
+
+// handleConfirmModalUpdate handles burst confirmation modal updates.
+func (i *Intent) handleConfirmModalUpdate(msg tea.Msg) tea.Cmd {
+	if i.confirmModal == nil || !i.confirmModal.IsVisible() {
+		return nil
+	}
+	cmd, confirmed := i.confirmModal.Update(msg)
+	if !i.confirmModal.IsVisible() {
+		if confirmed && i.selectedBurst != nil {
+			return tea.Batch(cmd, i.confirmBurst())
+		}
+		i.confirmModal = nil
+		if i.selectedBurst != nil {
+			// Show detail modal and return noopCmd to prevent message propagation.
+			// The message (e.g., Esc) was already consumed by the confirm modal.
+			i.showBurstDetailModal(i.selectedBurst)
+			return noopCmd
+		}
+		return noopCmd
+	}
+	return cmd
+}
+
+// handleSuggestionModalUpdate handles suggestion review modal updates.
+func (i *Intent) handleSuggestionModalUpdate(msg tea.Msg) tea.Cmd {
+	if i.suggestionModal == nil || !i.suggestionModal.IsVisible() {
+		return nil
+	}
+
+	// Handle accept key specially.
+	if keyMsg, ok := msg.(tea.KeyMsg); ok && keyMsg.String() == "a" {
+		return i.handleSuggestionAccept(msg)
+	}
+
+	// Handle other keys normally.
+	_, cmd := i.suggestionModal.Update(msg)
+	if !i.suggestionModal.IsVisible() {
+		return i.handleSuggestionModalClosed(cmd)
+	}
+	return cmd
+}
+
+// handleSuggestionAccept handles accepting a suggestion via 'a' key.
+func (i *Intent) handleSuggestionAccept(msg tea.Msg) tea.Cmd {
+	currentSuggestion := i.suggestionModal.GetCurrentSuggestion()
+	if currentSuggestion == nil {
+		return noopCmd
+	}
+
+	cmd := i.saveAndExtractBurst(*currentSuggestion)
+	i.suggestionModal.Update(msg)
+
+	if !i.suggestionModal.IsVisible() {
+		i.clearSuggestionState()
+		i.state = StateList
+		i.transitionToScreen(burstscreens.NewBurstListScreen(i.filteredBursts))
+	}
+	return cmd
+}
+
+// handleSuggestionModalClosed handles suggestion modal closure.
+func (i *Intent) handleSuggestionModalClosed(cmd tea.Cmd) tea.Cmd {
+	accepted := i.suggestionModal.GetAcceptedSuggestions()
+	cancelled := i.suggestionModal.GetAction() == burstmodals.SuggestionActionCancel
+	i.clearSuggestionState()
+
+	if len(accepted) > 0 && !cancelled {
+		completeMsg := SuggestionReviewCompleteMsg{
+			AcceptedSuggestions: accepted,
+			Cancelled:           false,
+		}
+		return tea.Batch(cmd, func() tea.Msg { return completeMsg })
+	}
+
+	i.state = StateList
+	i.transitionToScreen(burstscreens.NewBurstListScreen(i.filteredBursts))
+	return noopCmd
+}
+
+// handleDetailModalUpdate handles detail modal updates with keyboard shortcuts.
+func (i *Intent) handleDetailModalUpdate(msg tea.Msg) tea.Cmd {
+	if i.detailModal == nil || !i.detailModal.IsVisible() {
+		return nil
+	}
+
+	if keyMsg, ok := msg.(tea.KeyMsg); ok {
+		if cmd := i.handleDetailModalKeypress(keyMsg); cmd != nil {
+			return cmd
+		}
+	}
+
+	_, cmd := i.detailModal.Update(msg)
+	return cmd
+}
+
+// handleDetailModalKeypress handles keyboard shortcuts in detail modal.
+func (i *Intent) handleDetailModalKeypress(keyMsg tea.KeyMsg) tea.Cmd {
+	switch keyMsg.String() {
+	case "v":
+		i.detailModal.Hide()
+		return i.showBurstEventsModal()
+	case "f":
+		i.detailModal.Hide()
+		return i.showBurstFactsModal()
+	case "e":
+		i.detailModal.Hide()
+		return i.openEditModal(i.selectedBurst)
+	case "d":
+		i.detailModal.Hide()
+		return i.openDeleteModal(i.selectedBurst)
+	case "c":
+		i.detailModal.Hide()
+		return i.showConfirmBurstModal()
+	}
+
+	if keyMsg.Type == tea.KeyEsc || keyMsg.Type == tea.KeyEnter {
+		i.detailModal = nil
+		return noopCmd
+	}
+	return nil
+}
+
+// showConfirmBurstModal shows the confirmation modal for the selected burst.
+func (i *Intent) showConfirmBurstModal() tea.Cmd {
+	existingFacts := i.loadBurstFacts(i.selectedBurst)
+	if len(existingFacts) > 0 {
+		i.confirmModal = feedback.NewConfirmModal(
+			"Re-extract Facts",
+			fmt.Sprintf("This burst already has %d facts. Re-extract and add more?", len(existingFacts)),
+		).WithVariant(feedback.ConfirmDefault)
+	} else {
+		i.confirmModal = feedback.NewConfirmModal(
+			"Confirm Burst",
+			"Mark this burst as confirmed and extract facts?",
+		).WithVariant(feedback.ConfirmDefault)
+	}
+	return i.confirmModal.Init()
+}
+
+// handleEventsModalUpdate handles events modal updates.
+func (i *Intent) handleEventsModalUpdate(msg tea.Msg) tea.Cmd {
+	if i.eventsModal == nil || !i.eventsModal.IsVisible() {
+		return nil
+	}
+	if keyMsg, ok := msg.(tea.KeyMsg); ok {
+		if keyMsg.Type == tea.KeyEsc || keyMsg.Type == tea.KeyEnter {
+			i.eventsModal = nil
+			if i.selectedBurst != nil {
+				return i.showBurstDetailModal(i.selectedBurst)
+			}
+			return noopCmd
+		}
+	}
+	_, cmd := i.eventsModal.Update(msg)
+	return cmd
+}
+
+// handleFactsModalUpdate handles facts modal updates.
+func (i *Intent) handleFactsModalUpdate(msg tea.Msg) tea.Cmd {
+	if i.factsModal == nil || !i.factsModal.IsVisible() {
+		return nil
+	}
+	if keyMsg, ok := msg.(tea.KeyMsg); ok {
+		if keyMsg.Type == tea.KeyEsc || keyMsg.Type == tea.KeyEnter {
+			i.factsModal = nil
+			if i.selectedBurst != nil {
+				return i.showBurstDetailModal(i.selectedBurst)
+			}
+			return noopCmd
+		}
+	}
+	_, cmd := i.factsModal.Update(msg)
+	return cmd
+}
+
+// handleEditModalUpdate handles edit modal updates.
+func (i *Intent) handleEditModalUpdate(msg tea.Msg) tea.Cmd {
+	if i.editModal == nil || !i.editModal.IsVisible() {
+		return nil
+	}
+	cmd, completed, formData := i.editModal.Update(msg)
+	if !i.editModal.IsVisible() {
+		if completed && formData != nil && i.selectedBurst != nil {
+			editMsg := EditBurstMsg{
+				BurstID:     i.selectedBurst.ID,
+				Name:        formData.Name,
+				Description: formData.Description,
+			}
+			i.editModal = nil
+			return tea.Batch(cmd, func() tea.Msg { return editMsg })
+		}
+		i.editModal = nil
+		i.state = StateList
+		i.transitionToScreen(burstscreens.NewBurstListScreen(i.filteredBursts))
+		return noopCmd
+	}
+	return cmd
+}
+
+// =============================================================================
+// Keyboard Shortcut Handler
+// =============================================================================
+
+// handleKeyShortcuts handles keyboard shortcuts when no modal is active.
+func (i *Intent) handleKeyShortcuts(keyMsg tea.KeyMsg) tea.Cmd {
+	if keyMsg.String() == "s" {
+		// Trigger burst suggestion detection.
+		i.state = StateSuggesting
+		return i.startBurstDetection()
+	}
+	return nil
+}
+
+// =============================================================================
+// Message Handlers
+// =============================================================================
+
+// handleFactExtractionComplete handles the FactExtractionCompleteMsg.
+func (i *Intent) handleFactExtractionComplete(msg FactExtractionCompleteMsg) tea.Cmd {
+	i.extractingFacts = false
+	i.loadingModal = nil
+
+	if msg.Error != nil {
+		// Silently ignore cancelled operations - user already knows they cancelled.
+		if errors.Is(msg.Error, context.Canceled) {
+			i.state = StateList
+			return nil
+		}
+		i.errorModal = feedback.NewErrorModal("Extraction Failed", msg.Error.Error())
+		// Only transition to list if suggestion modal is not visible.
+		// User may still be reviewing remaining suggestions.
+		if i.suggestionModal == nil || !i.suggestionModal.IsVisible() {
+			i.state = StateList
+		}
+		return nil
+	}
+
+	i.extractedFactsCount = len(msg.Facts)
+
+	// If the suggestion modal is still visible (user reviewing remaining suggestions),
+	// do NOT change state or transition screens. Let user continue reviewing.
+	if i.suggestionModal != nil && i.suggestionModal.IsVisible() {
+		return nil
+	}
+
+	i.state = StateList
+
+	// Only show detail modal if a burst was selected (e.g., from confirm flow).
+	// When coming from suggestion acceptance flow, selectedBurst is nil,
+	// so we stay on the list view showing the newly created bursts.
+	if i.selectedBurst != nil {
+		return i.showBurstDetailModal(i.selectedBurst)
+	}
+
+	// Refresh list screen to ensure visual update after extraction completes.
+	i.transitionToScreen(burstscreens.NewBurstListScreen(i.filteredBursts))
+	return nil
+}
+
+// handleEditBurstMsg handles the EditBurstMsg sent by the edit modal.
+func (i *Intent) handleEditBurstMsg(msg EditBurstMsg) tea.Cmd {
+	if i.selectedBurst == nil || i.selectedBurst.ID != msg.BurstID {
+		// Burst mismatch or nil - show error.
+		i.errorModal = feedback.NewErrorModal("Edit Failed", "Burst not found")
+		i.state = StateDetail
+		return nil
+	}
+
+	// Validate the name is not empty.
+	if msg.Name == "" {
+		// Name is required - keep modal open and show error.
+		// Re-open the edit modal with an error indication.
+		i.state = StateEdit
+		return nil
+	}
+
+	// Clear the edit modal (it's already been closed by handleModalUpdates or test).
+	i.editModal = nil
+
+	// Save original values in case we need to rollback.
+	originalName := i.selectedBurst.Name
+	originalDescription := i.selectedBurst.Description
+
+	// Update burst with new values.
+	i.selectedBurst.Name = msg.Name
+	i.selectedBurst.Description = msg.Description
+
+	// Save to repository if available.
+	if i.context.BurstRepository != nil {
+		ctx := i.getContext()
+		err := i.context.BurstRepository.Update(ctx, i.selectedBurst)
+		if err != nil {
+			// Rollback in-memory changes on failure.
+			i.selectedBurst.Name = originalName
+			i.selectedBurst.Description = originalDescription
+			i.editError = err
+			i.errorModal = feedback.NewErrorModal("Update Failed", err.Error())
+			return nil
+		}
+	}
+
+	// Return to detail modal showing updated burst.
+	return i.showBurstDetailModal(i.selectedBurst)
+}
+
+// handleBurstEventsLoaded handles the BurstEventsLoadedMsg.
+func (i *Intent) handleBurstEventsLoaded(msg BurstEventsLoadedMsg) tea.Cmd {
+	i.loadingEvents = false
+	if !i.handleLoadError(msg.Error, "Load Events Failed") {
+		return nil
+	}
+	i.eventsModal = burstmodals.NewBurstEventsModal(
+		i.selectedBurst.ID, i.selectedBurst.Name, msg.Events, i.Theme())
+	i.showModalWithDimensions(i.eventsModal)
+	return nil
+}
+
+// handleBurstFactsLoaded handles the BurstFactsLoadedMsg.
+func (i *Intent) handleBurstFactsLoaded(msg BurstFactsLoadedMsg) tea.Cmd {
+	i.loadingFacts = false
+	if !i.handleLoadError(msg.Error, "Load Facts Failed") {
+		return nil
+	}
+	i.factsModal = burstmodals.NewBurstFactsModal(
+		i.selectedBurst.ID, i.selectedBurst.Name, msg.Facts, i.Theme())
+	i.showModalWithDimensions(i.factsModal)
+	return nil
+}
+
+// handleLoadError handles load errors and returns false if processing should stop.
+func (i *Intent) handleLoadError(err error, title string) bool {
+	if err != nil {
+		i.ShowErrorModal(title, err.Error())
+		return false
+	}
+	return i.selectedBurst != nil
+}
+
+// handleBurstSuggestionsLoaded handles the BurstSuggestionsLoadedMsg.
+func (i *Intent) handleBurstSuggestionsLoaded(msg BurstSuggestionsLoadedMsg) tea.Cmd {
+	i.suggestionsLoading = false
+	i.loadingModal = nil
+
+	if msg.Error != nil {
+		// Silently ignore cancelled operations - user already knows they cancelled.
+		if errors.Is(msg.Error, context.Canceled) {
+			i.state = StateList
+			return nil
+		}
+		i.suggestionsError = msg.Error
+		i.ShowErrorModal("Burst Detection Failed", msg.Error.Error())
+		i.state = StateList
+		return nil
+	}
+
+	if len(msg.Suggestions) == 0 {
+		// No suggestions found - show message and return to list.
+		i.ShowErrorModal("No Suggestions Found",
+			"No suggestions were generated from your events. "+
+				"Try adding more events or adjusting detection settings.")
+		i.state = StateList
+		return nil
+	}
+
+	// Show suggestion review modal and update state.
+	// Clear selectedBurst since we're entering suggestion review mode, not viewing a specific burst.
+	// This prevents handleFactExtractionComplete from showing the detail modal for an old burst.
+	i.selectedBurst = nil
+	i.suggestionModal = burstmodals.NewSuggestionReviewModal(msg.Suggestions, i.Theme())
+	width, height := i.getTerminalDimensions()
+	i.suggestionModal.SetDimensions(width, height)
+	i.suggestionModal.Show()
+	i.state = StateSuggestionReview
+	return nil
+}
+
+// handleSuggestionReviewComplete handles the SuggestionReviewCompleteMsg.
+func (i *Intent) handleSuggestionReviewComplete(msg SuggestionReviewCompleteMsg) tea.Cmd {
+	if !i.isInSuggestionState() {
+		return nil
+	}
+
+	if len(msg.AcceptedSuggestions) == 0 {
+		i.state = StateList
+		i.transitionToScreen(burstscreens.NewBurstListScreen(i.filteredBursts))
+		return nil
+	}
+
+	createdBursts := i.createBurstsFromSuggestions(msg.AcceptedSuggestions)
+	i.transitionToScreen(burstscreens.NewBurstListScreen(i.filteredBursts))
+
+	return i.startFactExtractionForBursts(createdBursts)
 }
